Update README.md
Emotion Detection Project
This project detects human emotions using a trained model and a user-friendly interface. The model was trained with Teachable Machine, and the project was developed with

HTML and CSS: For the layout and styling of the web interface.
JavaScript: For implementing logic and interaction with the trained model.

Features
Real-time emotion detection.
Easy-to-use interface.
Lightweight and efficient for web deployment.